Set Cucco AI behaviors and revenge trigger
The approach to the AI behavior is as follows:
If a Cucco is currently not in an angry state, and the player is invulnerable (i.e. gm is peaceful or player is in creative), the Cucco will not set a revenge target. Changing either of those states to a valid attack value will not cause the Cuccos to become hostile.
If the player is not invulnerable and a passive Cucco is attacked, it becomes hostile, making nearby existing Cuccos hositle, and spawning in a random number of additional hostile Cuccos.
If the target dies, Cuccos de-aggro. If the target becomes invulnerable, Cuccos will disengage until either their aggro timer runs out, or the target becomes vulnerable to damage again.
Aggressive state is persistent on world and game restart. Cuccos have to run their cycle all the way through.
If you are Harvy, may a higher being somewhere have mercy on you. The Cuccos will not show any