Added Cucco registry and rendering
Cucco is officially part of ZSS. Round of applause.
TODO come back to the model file and clean it up. It was just dropped into the project.
Also, blame Intellij for optimizing the imports in ZSSEntities. Imo it looks a little cleaner though. As long as you don't stare at the config entries too long. *shudders*