Cucco bug fixes Round 2
Added timer parameter to anger setter so that it would be easier to sync Cucco anger timers in a centralized way. If the target is null, the timer argument is ignored.
entity.rand.nextInt(bound) returns an integer between 0 and bound, exclusive. There was a chance that the random integer would be 0, and the spawn cycle would not process because it was 0 on the next tick. Added 1 to RNGs so that the range became 1 to bound, inclusive.