Cucco bug fixes and balancing
Turns out a 0.5 speed modifier is waaayyyy too fast. Sorry about that one...
Also, players would have never seen that speed boost anyways, because I was never applying the modifier... also my bad.
When a Cucco is currently seeking a target, but is actually not angry anymore, we need to reset the targeting task.
One last thing. Bumped up the base health value to 6 hearts. We needed to pose an actual challenge to players early-game, and discourage harassing the Cuccos. They need to actually be formidable.